+++ NEWS.md
@@ -1,6 +1,30 @@
+# AMAPVox 0.12.0
+
+*2022-08-24*
+Added new function "run" that either runs AMAPVox in batch mode or launches GUI
+Deleted gui/run check.update option, assuming that version set to "latest" implies to check for updates
+Released on CRAN.
+
+# AMAPVox 0.11.0
+
+*2022-07-25*
+Function readVoxelSpace can read zipped voxel file.
+Function plot accepts voxel subset in y parameter.
+Function plantAreaDensity returns separate data table and accepts voxel subset. 
+Not released on CRAN.
+
+# AMAPVox 0.10.1
+
+*2022-04-21*
+Bugfix in "toRaster" function.
+Bugfix across several functions ("canopyHeight", "groundElevation", "plantAreaIndex") for merging/joinging data.table based on "on" instead of "key".
+Bugfix in "computeG" function that speed up calculation for large theta vector.
+Not released on CRAN.
+
+
 # AMAPVox 0.10.0
 
-*2022-04-19*
+*2022-04-20*
 Added new functions "plantAreaDensity", "plantAreaIndex".
 Added new function "toRaster".
 Released on CRAN.